delete: remove obsolete classes
