--- src/CuiLib/Checkers/FileSystemCheckers.cs
@@ -1,94 +1,0 @@
-﻿using System;
-using System.IO;
-using CuiLib.Options;
-
-namespace CuiLib.Checkers
-{
-    /// <summary>
-    /// 読み込むファイルを検証します。
-    /// </summary>
-    [Obsolete($"{nameof(ValueChecker.VerifySourceFile)}を代わりに使用してください")]
-    [Serializable]
-    public class SourceFileChecker : IValueChecker<FileInfo>
-    {
-        /// <summary>
-        /// <see cref="SourceFileChecker"/>の新しいインスタンスを初期化します。
-        /// </summary>
-        public SourceFileChecker()
-        {
-        }
-
-        /// <inheritdoc/>
-        public ValueCheckState CheckValue(FileInfo value)
-        {
-            ArgumentNullException.ThrowIfNull(value);
-
-            DirectoryInfo? directory = value.Directory;
-            if (directory is not null && !directory.Exists) return ValueCheckState.AsError($"ファイル'{value.Name}'のディレクトリが存在しません");
-            if (!value.Exists) return ValueCheckState.AsError($"ファイル'{value.Name}'が存在しません");
-            return ValueCheckState.Success;
-        }
-    }
-
-    /// <summary>
-    /// 出力ファイルを検証します。
-    /// </summary>
-    [Obsolete($"{nameof(ValueChecker.VerifyDestinationFile)}を代わりに使用してください")]
-    [Serializable]
-    public class DestinationFileChecker : IValueChecker<FileInfo>
-    {
-        /// <summary>
-        /// 存在しないディレクトリを許容するかどうかを表す値を取得または設定します。
-        /// </summary>
-        public bool AllowMissedDirectory { get; set; }
-
-        /// <summary>
-        /// 上書きを許容するかどうかを表す値を取得または設定します。
-        /// </summary>
-        public bool AllowOverwrite { get; set; }
-
-        /// <summary>
-        /// <see cref="DestinationFileChecker"/>の新しいインスタンスを初期化します。
-        /// </summary>
-        public DestinationFileChecker()
-        {
-            AllowMissedDirectory = true;
-            AllowOverwrite = true;
-        }
-
-        /// <inheritdoc/>
-        public ValueCheckState CheckValue(FileInfo value)
-        {
-            ArgumentNullException.ThrowIfNull(value);
-
-            DirectoryInfo? directory = value.Directory;
-            if (!AllowMissedDirectory && directory is not null && !directory.Exists) return ValueCheckState.AsError($"ファイル'{value.Name}'のディレクトリが存在しません");
-            if (!AllowOverwrite && value.Exists) return ValueCheckState.AsError($"ファイル'{value.Name}'が既に存在しています");
-            return ValueCheckState.Success;
-        }
-    }
-
-    /// <summary>
-    /// 読み込むディレクトリを検証します。
-    /// </summary>
-    [Obsolete($"{nameof(ValueChecker.VerifySourceDirectory)}を代わりに使用してください")]
-    [Serializable]
-    public class SourceDirectoryChecker : IValueChecker<DirectoryInfo>
-    {
-        /// <summary>
-        /// <see cref="SourceDirectoryChecker"/>の新しいインスタンスを初期化します。
-        /// </summary>
-        public SourceDirectoryChecker()
-        {
-        }
-
-        /// <inheritdoc/>
-        public ValueCheckState CheckValue(DirectoryInfo value)
-        {
-            ArgumentNullException.ThrowIfNull(value);
-
-            if (!value.Exists) return ValueCheckState.AsError($"ディレクトリ'{value.Name}'が存在しません");
-            return ValueCheckState.Success;
-        }
-    }
-}
